+++ test/test_helper.exs
@@ -1,1 +1,7 @@
-ExUnit.start()
+children = [
+  {Task.Supervisor, name: ExZample.TestTaskSupervisor}
+]
+
+Supervisor.start_link(children, strategy: :one_for_one)
+
+ExUnit.start(exclude: [:skip])